==> ds_competition/3.logistic_regression.ipynb <==
# 载入所需要的库
import numpy as np
import pandas as pd
import tensorflow as tf
# 载入数据
X_test = pd.read_pickle('../input/X_test.pkl')
X_train = pd.read_pickle('../input/X_train.pkl')
y_train = pd.read_pickle('../input/y_train.pkl')

from sklearn.preprocessing import minmax_scale as mm
X_test = mm(X_test)
X_train= mm(X_train)

## 逻辑回归

import warnings 
warnings.filterwarnings("ignore")
from sklearn.linear_model import LogisticRegression
from sklearn.model_selection import GridSearchCV

# 一个训练优化模型的函数
def train_optimal_classifier(clf,param):
    grid_search = GridSearchCV(
        clf,
        param_grid=param,
        cv=5,
        verbose=0,
        scoring='f1_weighted',
        error_score=0,
        refit=True ,
        n_jobs=-1
        )
    grid_search.fit(X_train, y_train)
    print("Best parameters")
    print(grid_search.best_params_)
    print('best F1: ')
    print(grid_search.best_score_)
    return (grid_search.best_estimator_)


LR = LogisticRegression(class_weight='balanced',random_state=33,n_jobs=-1,solver='lbfgs')

param_LR = {  "C":[2e4,5e4,1e5,2e5,4e5,8e5],
             }
LR_best = train_optimal_classifier(LR,param_LR)

## Bagging 逻辑回归

from sklearn.ensemble import BaggingClassifier
BC = BaggingClassifier(LR_best)
param = {}
SVM_best = train_optimal_classifier(BC,param)

LR_best.fit(X_train, y_train)
y_pred = LR_best.predict()

data1 =  pd.read_csv("../input/studentID_test.csv",encoding='utf-8')

data2 = pd.DataFrame(y_pred)
data = pd.concat([data1,data2],axis =1)
data.to_csv('../data1.csv')
