==> ds_competition/5.neural_networks.ipynb <==
# 必要的库
import tensorflow as tf
import numpy as np
import pandas as pd

#加载数据
TRAINING = pd.read_csv("../input/train.csv")
TESTING = pd.read_csv("../input/test.csv")
# 将sub 数据换成０，１，２，３
sub_map = {0:0,1000:1,1500:2,2000:3}
def change(c): return sub_map[c]
TRAINING['sub'] = TRAINING['sub'].apply(change)

TESTING.shape

# 将数据转换成tensor格式方便TF的计算
feature_list = TESTING.columns.values.tolist()
def input_fn(df):
    feature_cols  = {k: tf.constant(df[k].values)
                     for k in feature_list}
    label = tf.constant(df['sub'].values)
    return feature_cols, label

def input_fn2(df):
    feature_cols  = {k: tf.constant(df[k].values)
                     for k in feature_list}
    return feature_cols


# Specify that all features have real-value data
rank = tf.contrib.layers.real_valued_column("rank")
card01 = tf.contrib.layers.real_valued_column("card01")
card02 = tf.contrib.layers.real_valued_column("card02")
card03 = tf.contrib.layers.real_valued_column("card03")
card04 = tf.contrib.layers.real_valued_column("card04")
card05 = tf.contrib.layers.real_valued_column("card05")
card06 = tf.contrib.layers.real_valued_column("card06")
card07 = tf.contrib.layers.real_valued_column("card07")
card08 = tf.contrib.layers.real_valued_column("card08")
card09 = tf.contrib.layers.real_valued_column("card09")
card10 = tf.contrib.layers.real_valued_column("card10")
card11 = tf.contrib.layers.real_valued_column("card11")
card12 = tf.contrib.layers.real_valued_column("card12")

feature_columns = [rank,card01,card02,card03,card04,card05,card06,card07,card08,card09,card10,card11,card12]
# 创建一个基本的神经网络
classifier = tf.contrib.learn.DNNClassifier(feature_columns=feature_columns,
                                            hidden_units=[13, 26, 13],
                                            n_classes=4,
                                            model_dir="/tmp/sub_model")

classifier.fit(input_fn=lambda: input_fn(TRAINING),steps=20000)

y = classifier.predict(input_fn=lambda: input_fn2(TESTING))
type(y)

listy = [x for x in y]

listy
